Switch to wvengen/proguard-maven-plugin plugin.
The original proguard-maven-plugin by pyx4me has not been maintained for nearly ten years and
some of its dependencies semm to have been removed from the central maven repository. The
proguard-maven-plugin by wvengen is a successor of the plugin by pyx4me. It is not under active
development but seems to be still maintained.